[ADD] 23/09/17 noticeBoard 생성, 리스트 조회 진행중
공지 게시글 등록 성공
공지 리스트 조회 진행중
 - 값은 받아오나 화면에 나오지 않음.

+++ src/main/java/com/kh/bluewave/noticeBoard/domain/NoticeBoard.java
@@ -10,7 +10,7 @@ public class NoticeBoard {
 	private String noticeFileName;
 	private String noticeFileRename;
 	private String noticeFilePath;
-	private int noticeFileLength;
+	private long noticeFileLength;
 	private int noticeViewCount;
 	private Timestamp noticeCreateDate;
 	private Timestamp noticeUpdateDate;
@@ -58,10 +58,10 @@ public class NoticeBoard {
 	public void setNoticeFilePath(String noticeFilePath) {
 		this.noticeFilePath = noticeFilePath;
 	}
-	public int getNoticeFileLength() {
+	public long getNoticeFileLength() {
 		return noticeFileLength;
 	}
-	public void setNoticeFileLength(int noticeFileLength) {
+	public void setNoticeFileLength(long noticeFileLength) {
 		this.noticeFileLength = noticeFileLength;
 	}
 	public int getNoticeViewCount() {
